move scheduler config out of componentconfig
