+++ hack/.golint_failures
@@ -325,6 +325,9 @@ pkg/registry/storage/storageclass
 pkg/registry/storage/storageclass/storage
 pkg/routes
 pkg/scheduler/api
+pkg/scheduler/api/v1
+pkg/scheduler/apis/schedulerconfig
+pkg/scheduler/apis/schedulerconfig/v1alpha1
 pkg/security/apparmor
 pkg/security/podsecuritypolicy
 pkg/security/podsecuritypolicy/group
