+++ pkg/apis/componentconfig/v1alpha1/register.go
@@ -43,8 +43,6 @@ func init() {
 }
 
 func addKnownTypes(scheme *runtime.Scheme) error {
-	scheme.AddKnownTypes(SchemeGroupVersion,
-		&KubeSchedulerConfiguration{},
-	)
+	scheme.AddKnownTypes(SchemeGroupVersion)
 	return nil
 }
