+++ pkg/apis/componentconfig/v1alpha1/types.go
@@ -20,94 +20,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-// ClientConnectionConfiguration contains details for constructing a client.
-type ClientConnectionConfiguration struct {
-	// kubeConfigFile is the path to a kubeconfig file.
-	KubeConfigFile string `json:"kubeconfig"`
-	// acceptContentTypes defines the Accept header sent by clients when connecting to a server, overriding the
-	// default value of 'application/json'. This field will control all connections to the server used by a particular
-	// client.
-	AcceptContentTypes string `json:"acceptContentTypes"`
-	// contentType is the content type used when sending data to the server from this client.
-	ContentType string `json:"contentType"`
-	// cps controls the number of queries per second allowed for this connection.
-	QPS float32 `json:"qps"`
-	// burst allows extra queries to accumulate when a client is exceeding its rate.
-	Burst int32 `json:"burst"`
-}
-
-// SchedulerPolicySource configures a means to obtain a scheduler Policy. One
-// source field must be specified, and source fields are mutually exclusive.
-type SchedulerPolicySource struct {
-	// File is a file policy source.
-	File *SchedulerPolicyFileSource `json:"file,omitempty"`
-	// ConfigMap is a config map policy source.
-	ConfigMap *SchedulerPolicyConfigMapSource `json:"configMap,omitempty"`
-}
-
-// SchedulerPolicyFileSource is a policy serialized to disk and accessed via
-// path.
-type SchedulerPolicyFileSource struct {
-	// Path is the location of a serialized policy.
-	Path string `json:"path"`
-}
-
-// SchedulerPolicyConfigMapSource is a policy serialized into a config map value
-// under the SchedulerPolicyConfigMapKey key.
-type SchedulerPolicyConfigMapSource struct {
-	// Namespace is the namespace of the policy config map.
-	Namespace string `json:"namespace"`
-	// Name is the name of hte policy config map.
-	Name string `json:"name"`
-}
-
-// SchedulerAlgorithmSource is the source of a scheduler algorithm. One source
-// field must be specified, and source fields are mutually exclusive.
-type SchedulerAlgorithmSource struct {
-	// Policy is a policy based algorithm source.
-	Policy *SchedulerPolicySource `json:"policy,omitempty"`
-	// Provider is the name of a scheduling algorithm provider to use.
-	Provider *string `json:"provider,omitempty"`
-}
-
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-
-type KubeSchedulerConfiguration struct {
-	metav1.TypeMeta `json:",inline"`
-
-	// SchedulerName is name of the scheduler, used to select which pods
-	// will be processed by this scheduler, based on pod's "spec.SchedulerName".
-	SchedulerName string `json:"schedulerName"`
-	// AlgorithmSource specifies the scheduler algorithm source.
-	AlgorithmSource SchedulerAlgorithmSource `json:"algorithmSource"`
-	// RequiredDuringScheduling affinity is not symmetric, but there is an implicit PreferredDuringScheduling affinity rule
-	// corresponding to every RequiredDuringScheduling affinity rule.
-	// HardPodAffinitySymmetricWeight represents the weight of implicit PreferredDuringScheduling affinity rule, in the range 0-100.
-	HardPodAffinitySymmetricWeight int32 `json:"hardPodAffinitySymmetricWeight"`
-
-	// LeaderElection defines the configuration of leader election client.
-	LeaderElection KubeSchedulerLeaderElectionConfiguration `json:"leaderElection"`
-
-	// ClientConnection specifies the kubeconfig file and client connection
-	// settings for the proxy server to use when communicating with the apiserver.
-	ClientConnection ClientConnectionConfiguration `json:"clientConnection"`
-	// HealthzBindAddress is the IP address and port for the health check server to serve on,
-	// defaulting to 0.0.0.0:10251
-	HealthzBindAddress string `json:"healthzBindAddress"`
-	// MetricsBindAddress is the IP address and port for the metrics server to
-	// serve on, defaulting to 0.0.0.0:10251.
-	MetricsBindAddress string `json:"metricsBindAddress"`
-	// EnableProfiling enables profiling via web interface on /debug/pprof
-	// handler. Profiling handlers will be handled by metrics server.
-	EnableProfiling bool `json:"enableProfiling"`
-	// EnableContentionProfiling enables lock contention profiling, if
-	// EnableProfiling is true.
-	EnableContentionProfiling bool `json:"enableContentionProfiling"`
-
-	// Indicate the "all topologies" set for empty topologyKey when it's used for PreferredDuringScheduling pod anti-affinity.
-	FailureDomains string `json:"failureDomains"`
-}
-
 // LeaderElectionConfiguration defines the configuration of leader election
 // clients for components that can run with leader election enabled.
 type LeaderElectionConfiguration struct {
@@ -135,23 +47,3 @@ type LeaderElectionConfiguration struct {
 	// during leader election cycles.
 	ResourceLock string `json:"resourceLock"`
 }
-
-// KubeSchedulerLeaderElectionConfiguration expands LeaderElectionConfiguration
-// to include scheduler specific configuration.
-type KubeSchedulerLeaderElectionConfiguration struct {
-	LeaderElectionConfiguration `json:",inline"`
-	// LockObjectNamespace defines the namespace of the lock object
-	LockObjectNamespace string `json:"lockObjectNamespace"`
-	// LockObjectName defines the lock object name
-	LockObjectName string `json:"lockObjectName"`
-}
-
-const (
-	// "kube-system" is the default scheduler lock object namespace
-	SchedulerDefaultLockObjectNamespace string = "kube-system"
-
-	// "kube-scheduler" is the default scheduler lock object name
-	SchedulerDefaultLockObjectName = "kube-scheduler"
-
-	SchedulerDefaultProviderName = "DefaultProvider"
-)
